+++ common/pal.h
@@ -61,7 +61,7 @@ bool pal_load_IPMB_config(void);
 void pal_set_sensor_poll_interval(int *interval_ms);
 
 // sensor accessible
-uint8_t pal_load_sdr_table(void);
+uint16_t pal_load_sdr_table(void);
 bool pal_load_snr_config(void);
 void pal_fix_fullSDR_table(void);
 void pal_fix_Snrconfig(void);
